+++ ahem/loader.py
@@ -1,6 +1,4 @@
-"""
-Main notification dispatcher.
-"""
+
 import inspect
 from importlib import import_module
